+++ source/encoder_video_h264_libx264/changelog.md
@@ -1,3 +1,6 @@
+**<span style="color:#56adda">0.0.3</span>**
+- Add HW accelerated decoding option
+
 **<span style="color:#56adda">0.0.2</span>**
 - Add ability to set the output container
 - Add icon